backend/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-auth==2.20.0